+++ apps/miranum-modeler-dmn-webview/vite.config.ts
@@ -1,11 +1,10 @@
 /// <reference types="vitest" />
 import { defineConfig } from "vite";
-
 import { viteStaticCopy } from "vite-plugin-static-copy";
 import path from "path";
 
 export default defineConfig({
-    cacheDir: "../../node_modules/.vite/miranum-bpmn-modeler-webview",
+    cacheDir: "../../node_modules/.vite/miranum-modeler-dmn-webview",
 
     resolve: {
         alias: [
@@ -23,13 +22,6 @@ export default defineConfig({
                     "../../libs/vscode/shared/miranum-modeler/src",
                 ),
             },
-            {
-                find: "@miranum-ide/miranum-create-append-c7-element-templates",
-                replacement: path.resolve(
-                    __dirname,
-                    "../../libs/miranum-create-append-c7-element-templates/src",
-                ),
-            },
         ],
     },
 
@@ -37,11 +29,11 @@ export default defineConfig({
         viteStaticCopy({
             targets: [
                 {
-                    src: "../../node_modules/camunda-bpmn-js/dist/assets/bpmn-font/css/**",
+                    src: "../../node_modules/dmn-js/dist/assets/dmn-font/css/**",
                     dest: "css/",
                 },
                 {
-                    src: "../../node_modules/camunda-bpmn-js/dist/assets/bpmn-font/font/**",
+                    src: "../../node_modules/dmn-js/dist/assets/dmn-font/font/**",
                     dest: "font/",
                 },
             ],