+++ apps/miranum-extension-pack/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* **\[Miranum Config Editor\]** Add new plugin for editing config files (#412)
+
 ### Changed
 
 * Update dependencies and migrate to Nx Version 17 (#418)